# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DistrictAssist AI",
    "version": "1.0.0",
    "description": "API for managing student data, imports, and support plans."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/v1/districts": {
      "post": {
        "summary": "Create a district",
        "tags": ["Districts"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "stateCode": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 2
                  },
                  "clerkOrganizationId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 255
                  }
                },
                "required": ["name", "stateCode"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
    "/api/v1/districts/{districtId}/students": {
      "post": {
        "summary": "Create a student",
        "tags": ["Students"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "districtId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "externalId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "lastName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "gradeLevel": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 12
                  },
                  "schoolName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "programStatus": {
                    "type": "string",
                    "enum": ["active", "monitoring", "inactive"]
                  }
                },
                "required": [
                  "externalId",
                  "firstName",
                  "lastName",
                  "gradeLevel",
                  "schoolName",
                  "programStatus"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "summary": "List students",
        "tags": ["Students"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/districts/{districtId}/students/{studentId}/support-plans": {
      "post": {
        "summary": "Create a support plan",
        "tags": ["Support Plans"],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "districtId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "required": true,
            "name": "studentId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goal": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "reviewDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["draft", "active", "completed", "cancelled"],
                    "default": "draft"
                  }
                },
                "required": ["goal", "startDate", "reviewDate"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      },
      "get": {
        "summary": "List support plans",
        "tags": ["Support Plans"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}